+++ src/main/java/io/github/forgestove/create_cyber_goggles/config/client/gui/entry/StringValueConfigEntry.java
@@ -1,0 +1,8 @@
+package io.github.forgestove.create_cyber_goggles.config.client.gui.entry;
+import io.github.forgestove.create_cyber_goggles.config.client.gui.ConfigCategoryTab;
+import io.github.forgestove.create_cyber_goggles.config.tree.ValueConfigNode;
+public final class StringValueConfigEntry<C> extends GenericValueConfigEntry<C, String> {
+	public StringValueConfigEntry(ConfigCategoryTab<C> tab, ValueConfigNode<C, String> node) {
+		super(tab, node, s -> s, s -> true);
+	}
+}