+++ src/main/java/io/github/forgestove/create_cyber_goggles/config/tree/ConfigNode.java
@@ -11,4 +11,7 @@ public interface ConfigNode<C> {
 	@Nullable Component validate(C config);
 	void copy(C from, C to);
 	void writeEditingToConfig(C config);
+	default String getPath() {
+		return "";
+	}
 }